feat(bex): fixed issue for Relayer Approval specific to using useTxn inside a hook where the write loading state is never lifted into CreatePageContent.
